requirements.txt
# FastAPI и веб-сервер
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Telegram Bot
aiogram==3.3.0

# Безопасность
pyjwt==2.8.0
cryptography==42.0.0

# Утилиты
python-dotenv==1.0.0
python-multipart==0.0.6
